+++ tests/htlc_harness/ether_harness.rs
@@ -75,7 +75,11 @@ pub fn ether_harness<D: Docker>(
         )
         .into(),
         params.htlc_wei_value,
+        EtherHtlc::deploy_tx_gas_limit().into(),
     );
+
+    let transaction_receipt = alice_client.receipt(tx_id);
+    log::debug!("used gas Ether deploy {:?}", transaction_receipt.gas_used);
 
     (
         alice,