+++ tests/rfc003_erc20_htlc.rs
@@ -43,7 +43,7 @@ fn given_erc20_token_should_deploy_erc20_htlc_and_fund_htlc() {
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
     // Fund erc20 htlc
-    client.sign_and_send(|nonce, gas_price| UnsignedTransaction {
+    let tx = client.sign_and_send(|nonce, gas_price| UnsignedTransaction {
         nonce,
         gas_price,
         gas_limit: U256::from(100_000),
@@ -58,6 +58,9 @@ fn given_erc20_token_should_deploy_erc20_htlc_and_fund_htlc() {
         ),
     });
 
+    let transaction_receipt = client.receipt(tx);
+    log::debug!("used gas ERC20 fund {:?}", transaction_receipt.gas_used);
+
     // Check htlc funding
     assert_eq!(
         client.token_balance_of(token_contract, htlc_address),
@@ -70,7 +73,12 @@ fn given_erc20_token_should_deploy_erc20_htlc_and_fund_htlc() {
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
     // Send correct secret to contract
-    client.send_data(htlc_address, Some(Bytes(SECRET.to_vec())));
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        Some(Bytes(SECRET.to_vec())),
+        Erc20Htlc::redeem_tx_gas_limit().into(),
+    );
+    log::debug!("used gas ERC20 redeem {:?}", transaction_receipt.gas_used);
 
     assert_eq!(
         client.token_balance_of(token_contract, htlc_address),
@@ -119,7 +127,11 @@ fn given_funded_erc20_htlc_when_redeemed_with_secret_then_tokens_are_transferred
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
     // Send correct secret to contract
-    let transaction_receipt = client.send_data(htlc_address, Some(Bytes(SECRET.to_vec())));
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        Some(Bytes(SECRET.to_vec())),
+        U256::from(Erc20Htlc::redeem_tx_gas_limit()),
+    );
     log::debug!("used gas ERC20 redeemed {:?}", transaction_receipt.gas_used);
 
     assert_eq!(
@@ -171,7 +183,8 @@ fn given_deployed_erc20_htlc_when_refunded_after_expiry_time_then_tokens_are_ref
 
     // Wait for the contract to expire
     sleep_until(harness_params.htlc_refund_timestamp);
-    let transaction_receipt = client.send_data(htlc_address, None);
+    let transaction_receipt =
+        client.send_data(htlc_address, None, Erc20Htlc::refund_tx_gas_limit().into());
     log::debug!("used gas ERC20 refund {:?}", transaction_receipt.gas_used);
 
     assert_eq!(
@@ -224,9 +237,10 @@ fn given_deployed_erc20_htlc_when_expiry_time_not_yet_reached_should_revert_tx_w
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
     // Don't wait for the timeout and don't send a secret
-    let transaction_receipt = client.send_data(htlc_address, None);
+    let transaction_receipt =
+        client.send_data(htlc_address, None, Erc20Htlc::refund_tx_gas_limit().into());
     log::debug!(
-        "used gas ERC20 too early {:?}",
+        "used gas ERC20 refund too early {:?}",
         transaction_receipt.gas_used
     );
 
@@ -282,7 +296,15 @@ fn given_not_enough_tokens_when_redeemed_token_balances_dont_change() {
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
     // Send correct secret to contract
-    client.send_data(htlc_address, Some(Bytes(SECRET.to_vec())));
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        Some(Bytes(SECRET.to_vec())),
+        Erc20Htlc::redeem_tx_gas_limit().into(),
+    );
+    log::debug!(
+        "used gas ERC20 redeemed not enough token {:?}",
+        transaction_receipt.gas_used
+    );
 
     assert_eq!(
         client.token_balance_of(token_contract, htlc_address),
@@ -319,7 +341,11 @@ fn given_htlc_and_redeem_should_emit_redeem_log_msg_with_secret() {
     });
 
     // Send correct secret to contract
-    let transaction_receipt = client.send_data(htlc_address, Some(Bytes(SECRET.to_vec())));
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        Some(Bytes(SECRET.to_vec())),
+        500_000.into(), // This is for test purposes only
+    );
     log::debug!("used gas ERC20 redeem {:?}", transaction_receipt.gas_used);
 
     // Should contain 2 logs: 1 for token transfer 1 for redeeming the htlc
@@ -365,7 +391,13 @@ fn given_htlc_and_refund_should_emit_refund_log_msg() {
     // Wait for the contract to expire
     sleep_until(harness_params.htlc_refund_timestamp);
     // Send correct secret to contract
-    let transaction_receipt = client.send_data(htlc_address, None);
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        None,
+        500_000.into(), // This is for test purposes only
+    );
+
+    log::debug!("used gas ERC20 refund {:?}", transaction_receipt.gas_used);
 
     // Should contain 2 logs: 1 for token transfer 1 for redeeming the htlc
     assert_that(&transaction_receipt.logs.len()).is_equal_to(2);
@@ -429,6 +461,12 @@ fn given_funded_erc20_htlc_when_redeemed_with_short_secret_should_revert_with_er
     let transaction_receipt = client.send_data(
         htlc_address,
         Some(Bytes(vec![1u8, 2u8, 3u8, 4u8, 6u8, 6u8, 7u8, 9u8, 10u8])),
+        Erc20Htlc::redeem_tx_gas_limit().into(),
+    );
+
+    log::debug!(
+        "used gas ERC20 redeem short secret {:?}",
+        transaction_receipt.gas_used
     );
 
     assert_eq!(
@@ -486,7 +524,13 @@ fn given_correct_zero_secret_htlc_should_redeem() {
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
     // Send short secret to contract
-    client.send_data(htlc_address, Some(Bytes(secret_vec)));
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        Some(Bytes(secret_vec)),
+        Erc20Htlc::redeem_tx_gas_limit().into(),
+    );
+
+    log::debug!("used gas ERC20 redeem {:?}", transaction_receipt.gas_used);
 
     assert_eq!(
         client.token_balance_of(token_contract, htlc_address),
@@ -549,6 +593,12 @@ fn given_short_zero_secret_htlc_should_revert_tx_with_error() {
             0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
             0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
         ])),
+        Erc20Htlc::redeem_tx_gas_limit().into(),
+    );
+
+    log::debug!(
+        "used gas ERC20 redeem zero secret {:?}",
+        transaction_receipt.gas_used
     );
 
     assert_eq!(
@@ -601,7 +651,11 @@ fn given_invalid_secret_htlc_should_revert_tx_with_error() {
     );
     assert_eq!(client.token_balance_of(token_contract, bob), U256::from(0));
 
-    let transaction_receipt = client.send_data(htlc_address, Some(Bytes(secret_vec)));
+    let transaction_receipt = client.send_data(
+        htlc_address,
+        Some(Bytes(secret_vec)),
+        500_000.into(), // This is for test purposes only
+    );
     log::debug!(
         "used gas ERC20 invalid secret {:?}",
         transaction_receipt.gas_used